ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.14.0
jiwer>=3.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
tensorflow==2.14.0
tensorflowjs>=4.10.0
Pillow>=10.0.0
pathlib2>=2.3.7
